Добавляет идемпотентность запросам

+++ Application/Features/Categories/Commands/RemoveCategory/RemoveCategoryHandler.cs
@@ -9,11 +9,12 @@ internal sealed class RemoveCategoryHandler(IDbContext context) : IRequestHandle
 {
     public async Task Handle(RemoveCategoryCommand request, CancellationToken cancellationToken)
     {
-        var category = await context.Categories.GetAsync(new GetCategoryByIdInputData
+        var category = await context.Categories.GetAsync(new GetCategoryByIdParameters
         {
             CategoryId = request.RouteDto.CategoryId,
             AsTracking = true,
         }, cancellationToken);
+        if (ReferenceEquals(category, default)) return;
         
         context.Categories.Remove(category);
